Move tests to the right files
